mount: Pass more arguments to mount command
When using the *fstab* argument of the *mount*` module, combined with
*state=mounted*, the `mount` call fails as it is not given enough
information. Explicitly calling the `mount` command with filesystem
type, options, block device, and mount point allow it to succeed no
matter what the value of the *fstab* argument is. In this way, the
*mount* module can be used to mount filesystems without updating
`/etc/fstab`, by passing *fstab=/dev/null*.

Signed-off-by: Dustin C. Hatch <admiralnemo@gmail.com>